Added convertToNativeSql option
This was causing us grief when using the sqlFile changeset feature and our file contained question mark characters ('?').   There doesn't seem to be a reasonable use case for parameterizing the sql in a file, so we made the SQLFileChange default to false/don't convert the sql to native.  When using a JDBC connection this avoids the nativeSql convertion of '?' to native format. eg. Oracle uses ':n' pattern where n is the ordinal parameter number.